+++ src/main/java/com/github/Naroru/JavaRushTelegramBot/command/CommandContainer.java
@@ -2,7 +2,6 @@ package com.github.Naroru.JavaRushTelegramBot.command;
 
 import com.github.Naroru.JavaRushTelegramBot.command.basicCommands.*;
 import com.github.Naroru.JavaRushTelegramBot.javarushclient.JavaRushGroupClient;
-import com.github.Naroru.JavaRushTelegramBot.javarushclient.JavaRushGroupClientImp;
 import com.github.Naroru.JavaRushTelegramBot.service.GroupSubsciptionService;
 import com.github.Naroru.JavaRushTelegramBot.service.SendMessageService;
 import com.github.Naroru.JavaRushTelegramBot.service.TelegramUserService;
@@ -29,6 +28,7 @@ public class CommandContainer {
        commands.put(CommandName.HELP.getCommandName(), new HelpCommand(sendMessageService));
        commands.put(CommandName.STAT.getCommandName(), new StatCommand(sendMessageService, telegramUserService));
        commands.put(CommandName.ADD_GROUP_SUB.getCommandName(), new AddGroupSubCommand(sendMessageService, javaRushGroupClient, groupSubsciptionService));
+       commands.put(CommandName.GET_GROUP_LIST.getCommandName(),new GetGroupListCommand(sendMessageService,telegramUserService));
 
        unknownCommand = new UnknowCommand(sendMessageService);
    }