+++ src/test/java/com/github/Naroru/JavaRushTelegramBot/command/CommandContainerTest.java
@@ -2,6 +2,7 @@ package com.github.Naroru.JavaRushTelegramBot.command;
 
 import com.github.Naroru.JavaRushTelegramBot.command.basicCommands.UnknowCommand;
 import com.github.Naroru.JavaRushTelegramBot.javarushclient.JavaRushGroupClient;
+import com.github.Naroru.JavaRushTelegramBot.service.GroupSubsciptionService;
 import com.github.Naroru.JavaRushTelegramBot.service.SendMessageService;
 import com.github.Naroru.JavaRushTelegramBot.service.TelegramUserService;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,12 +29,15 @@ class CommandContainerTest {
     @Mock
     private JavaRushGroupClient javaRushGroupClient;
 
+    @Mock
+    private GroupSubsciptionService groupSubsciptionService;
+
     private CommandContainer commandContainer;
 
     @BeforeEach
     public void init()
     {
-        commandContainer = new CommandContainer(sendMessageService, telegramUserService, javaRushGroupClient);
+        commandContainer = new CommandContainer(sendMessageService, telegramUserService, javaRushGroupClient,groupSubsciptionService);
     }
 
     @Test