+++ src/test/java/com/github/Naroru/JavaRushTelegramBot/command/basicCommands/AbstractCommandTest.java
@@ -25,8 +25,6 @@ import static org.mockito.Mockito.when;
 
      abstract Command getCommand();
 
-    @Mock
-    Update update;
 
     @Mock
     TelegramUserService telegramUserService;